requirements.txt
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0